app/requirements.txt
Django==3.1.7
gunicorn==20.1.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
requests==2.25.1